refactoring for testing
