+++ main.cpp
@@ -1,55 +1,6 @@
-#include <iostream>
-#include <string>
-#include <vector>
+#include "student_database.h"
 
-struct Student {
-  std::string name;
-  int age;
-  std::string major;
-  double gpa;
-};
 
-void addStudent(std::vector<Student> &database) {
-  Student student;
-  std::cout << "Введите имя студента: ";
-  std::cin >> student.name;
-  std::cout << "Введите возраст студента: ";
-  std::cin >> student.age;
-  std::cout << "Введите специальность студента: ";
-  std::cin >> student.major;
-  std::cout << "Введите средний балл студента: ";
-  std::cin >> student.gpa;
-
-  database.push_back(student);
-  std::cout << "Студент добавлен в базу данных.\n";
-}
-
-void print_sudent(const Student &student) {
-  std::cout << "Имя: " << student.name << "\n";
-  std::cout << "Возраст: " << student.age << "\n";
-  std::cout << "Специальность: " << student.major << "\n";
-  std::cout << "Средний балл: " << student.gpa << "\n\n";
-}
-
-void displayStudents(const std::vector<Student> &database) {
-  std::cout << "Список студентов:\n";
-  for (const Student &student : database) {
-    print_sudent(student);
-  }
-}
-
-void search_by_name_or_specialty(const std::vector<Student> &database) {
-  std::cout << "Введите имя студента или специальность: ";
-  std::string target;
-  std::cin >> target;
-  for (const Student &student : database) {
-    if (student.name == target || student.major == target) {
-      print_sudent(student);
-      return;
-    }
-  }
-  std::cout << "Нет студента с таким именем или специальностью\n";
-}
 
 int main() {
   std::vector<Student> database;
